requirements.txt
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.125.0
pandas==2.2.0
PyPDF2==3.0.1
openai==1.30.3
gspread==6.1.2
python-dotenv==0.15.0
lxml==5.2.2
selenium==4.22.0
fuzzywuzzy==0.18.0
Whoosh==2.7.4